regen all
