+++ project/jsonld/confident_schema.context.jsonld
@@ -1,5 +1,5 @@
 {
-   "_comments": "Auto generated from ConfIDent_schema.yaml by jsonldcontextgen.py version: 0.1.1\n    Generation date: 2022-10-17T12:54:48\n    Schema: confident_schema\n    metamodel version: 1.7.0\n    model version: 0.5.0\n    \n    id: https://raw.githubusercontent.com/TIBHannover/ConfIDent_schema/main/src/linkml/ConfIDent_schema.yaml\n    description: This is a schema for the ConfIDent project that describes the necessary metadata requirements of academic events and event series.\n    license: https://creativecommons.org/licenses/by/4.0/\n    ",
+   "_comments": "Auto generated from ConfIDent_schema.yaml by jsonldcontextgen.py version: 0.1.1\n    Generation date: 2022-10-17T22:58:16\n    Schema: confident_schema\n    metamodel version: 1.7.0\n    model version: 0.5.0\n    \n    id: https://raw.githubusercontent.com/TIBHannover/ConfIDent_schema/main/src/linkml/ConfIDent_schema.yaml\n    description: This is a schema for the ConfIDent project that describes the necessary metadata requirements of academic events and event series.\n    license: https://creativecommons.org/licenses/by/4.0/\n    ",
    "@context": {
       "BFO": {
          "@id": "http://purl.obolibrary.org/obo/BFO_",
